Changed: Inline common module functions into tools
The common module in llm-coding-tools-serdesai was only used by the
tools module. Inlined the helper functions directly into their
respective tool files to simplify the module structure:

- edit.rs: inlined error_to_serdes()
- glob.rs: inlined glob_output_to_return() and output_content()
- grep.rs: inlined grep_output_to_return()

Removed the now-unused common/ directory and mod common declaration.
