--- src/llm-coding-tools-serdesai/src/common/mod.rs
@@ -1,5 +1,0 @@
-//! Shared helpers for tool implementations.
-
-pub(crate) mod edit;
-pub(crate) mod glob;
-pub(crate) mod grep;
