+++ src/llm-coding-tools-serdesai/src/lib.rs
@@ -3,7 +3,6 @@
 
 pub mod agent_ext;
 pub mod agent_runtime;
-mod common;
 pub mod convert;
 pub mod task;
 pub mod tools;
