+++ src/llm-coding-tools-serdesai/src/tools/edit.rs
@@ -14,13 +14,13 @@ use llm_coding_tools_core::ToolContext;
 use llm_coding_tools_core::context::{PathMode, ToolPrompt};
 use llm_coding_tools_core::path::PathResolver;
 use llm_coding_tools_core::tool_metadata::edit as edit_meta;
-use llm_coding_tools_core::tools::edit_file;
+use llm_coding_tools_core::tools::{EditError, edit_file};
 use serde::Deserialize;
 use serdes_ai::tools::{
     RunContext, SchemaBuilder, Tool, ToolDefinition, ToolError, ToolResult, ToolReturn,
 };
 
-use crate::common::edit::error_to_serdes;
+use crate::convert::core_error_to_serdes;
 
 /// Internal args for JSON deserialization.
 #[derive(Debug, Deserialize)]
@@ -86,6 +86,38 @@ impl<R: PathResolver + Clone + Send + Sync, Deps: Send + Sync> Tool<Deps> for Ed
         .await;
 
         result.map(ToolReturn::text).map_err(error_to_serdes)
+    }
+}
+
+/// Convert [`EditError`] to serdesAI error.
+///
+/// Maps edit-specific errors to appropriate error types:
+/// - Validation errors: `NotFound`, `AmbiguousMatch`, `EmptyOldString`, `IdenticalStrings`
+/// - Execution errors: `Tool(ToolError)` (IO, path errors)
+fn error_to_serdes(err: EditError) -> ToolError {
+    match err {
+        EditError::NotFound => ToolError::validation_error(
+            edit_meta::NAME,
+            Some("old_string".to_string()),
+            "old_string not found in file content".to_string(),
+        ),
+        EditError::AmbiguousMatch => ToolError::validation_error(
+            edit_meta::NAME,
+            Some("old_string".to_string()),
+            "old_string found multiple times and requires more code context to uniquely identify the intended match"
+                .to_string(),
+        ),
+        EditError::EmptyOldString => ToolError::validation_error(
+            edit_meta::NAME,
+            Some("old_string".to_string()),
+            "old_string must not be empty".to_string(),
+        ),
+        EditError::IdenticalStrings => ToolError::validation_error(
+            edit_meta::NAME,
+            None,
+            "old_string and new_string must be different".to_string(),
+        ),
+        EditError::Tool(tool_err) => core_error_to_serdes(edit_meta::NAME, tool_err),
     }
 }
